Add "slimfast" command
The slimfast command allows to make the .docx more lightweight by
performing conversions in the attached pictures : quite often, some .emf
files are (uselessly) large. A first attempt is to convert them as svg,
which is sometimes enough to make the file much smaller. This is done
using "emf2svg-conv" from https://github.com/kakwa/libemf2svg . The
resulting svg quality looks good but it would require more tests.

If the resulting file is still large, the SVG is rasterized to JPG.
The embedded PNG files are also converted to JPG (of course, all
transparency info is lost, and more generally JPG is a lossy
compression, but overall it allows to make the .docx much smaller while
still keeping acceptable quality).